Bugfix for controlnet style transfer + regional.
In controlnet's hook, a set of 8 extra values are appended to the end of
the pos/neg context. This applies them to each region.

NOTE: Requires a PR merge on controlnet to function, otherwise this
shouldn't do anything (unless there's  some other context symmetry
breaking module of which I am unaware).