OVERHAUL: Promptcessor.
- Feature: Added new keyword, ADDUPE with optional number.
  This creates a placeholder region, which doesn't have its own context
(saves computation), but rather the context reference is duplicated in
attention and the mask is edited in latent.
  Use eg ADDUPE-2 to reference region which was 2 before current, or
ADDUPE1 for the first region, or ADDUPE0 for pure base.
  Former prompt pos + neg extensions, which had been disabled, have been
restored under this keyword, and also added to inpaint+.
- Feature: Added negative base parameter. Mostly applies to inpaint+, to
prevent errors and allow specifying prompt for non coloured regions, but
technically identical to base.
- Experimental: Negative prompt regions in latent mode. Added a setting
for it. Normally, uncond only gets one region in x_out. So instead it
appends the neg to pos, then in denoised nullifies the placeholder neg
entirely and negates x_out's negs regionally (which is what
combine_denoised does). This needs to be optimised, tested with batch,
and I'm not 100% sure it's even working correctly: the output looks
about right, pretty close to single uncond, but not identical.
- Enhancement: Debug mode has been moved fully to settings (stored when
the run begins).
- Enhancement: Added setting to force clear hooks due to some lora
forward cleanup mishaps.
- Refactor: Split off all text editing methods to promptcessor file.
Added key parameter and simplified the process somewhat, extra
functions.
- Refactor: Rewrote / rearranged several loose "self" functions as class
methods, inherited to script.

Prompt mode has not been enhanced, and tested minimally (attention looks
okay, latent might be broken).